Make nTox work on MinGW/Win32
