+++ configure.ac
@@ -115,8 +115,6 @@ AC_CANONICAL_HOST
 case $host_os in
     *mingw*)
         WIN32="yes"
-        AC_MSG_WARN([nTox is not supported on $host_os yet, disabling])
-        BUILD_NTOX="no"
         EXTRA_LT_LDFLAGS="$EXTRA_LT_LDFLAGS -no-undefined"
     ;;
     *solaris*)
@@ -202,9 +200,8 @@ AC_TYPE_UINT8_T
 
 # Checks for library functions.
 AC_FUNC_FORK
-AC_FUNC_MALLOC
 AC_FUNC_REALLOC
-AC_CHECK_FUNCS([gettimeofday memset socket strchr])
+AC_CHECK_FUNCS([gettimeofday memset socket strchr malloc])
 
 # pkg-config based tests
 PKG_PROG_PKG_CONFIG
@@ -271,22 +268,41 @@ if (test "x$BUILD_NTOX" = "xyes") && (test "x$NCURSES_FOUND" != "xyes"); then
             ]
         )
         if test "x$BUILD_NTOX" = "xyes"; then
-            AC_CHECK_LIB([ncurses], [clear],
-                [],
-                [
-                    unset ac_cv_lib_ncurses_clear
-                    AC_CHECK_LIB([ncurses], [clear],
-                        [],
-                        [
-                            AC_MSG_WARN([not building nTox client because required library ncurses was not found on your system])
-                            BUILD_NTOX="no"
-                        ],
-                        [
-                            -ltinfo
-                        ]
-                    )
-                ]
-            )
+            if test "x$WIN32" = "xyes"; then
+                AC_CHECK_LIB([pdcurses], [clear],
+                    [
+                        NCURSES_LIBS="-lpdcurses"
+                        AC_SUBST(NCURSES_LIBS)
+                    ],
+                    [
+                        AC_MSG_ERROR([required library pdcurses was not found on your system])
+                        BUILD_NTOX="no"
+                    ]
+                )
+            else
+                AC_CHECK_LIB([ncurses], [clear],
+                    [
+                        NCURSES_LIBS="-lncurses"
+                        AC_SUBST(NCURSES_LIBS)
+                    ],
+                    [
+                        unset ac_cv_lib_ncurses_clear
+                        AC_CHECK_LIB([ncurses], [clear],
+                            [
+                                NCURSES_LIBS="-lncurses -ltinfo"
+                                AC_SUBST(NCURSES_LIBS)
+                            ],
+                            [
+                                AC_MSG_WARN([not building nTox client because required library ncurses was not found on your system])
+                                BUILD_NTOX="no"
+                            ],
+                            [
+                                -ltinfo
+                            ]
+                        )
+                    ]
+                )
+            fi
         fi
     fi
 fi
