+++ testing/Makefile.inc
@@ -13,8 +13,8 @@ nTox_CFLAGS =           $(LIBSODIUM_CFLAGS) \
 nTox_LDADD =            $(LIBSODIUM_LDFLAGS) \
                         libtoxcore.la \
                         $(LIBSODIUM_LIBS) \
-                        $(NCURSES_LIBS)
-
+                        $(NCURSES_LIBS) \
+                        $(WINSOCK2_LIBS)
 endif
